+++ back-end/src/main/java/br/edu/com/tg/manager/application/service/ImportStudentsFromCsvService.java
@@ -5,10 +5,10 @@ import br.edu.com.tg.manager.core.entity.Course;
 import br.edu.com.tg.manager.core.entity.Student;
 import br.edu.com.tg.manager.core.entity.StudentGroup;
 import br.edu.com.tg.manager.core.exception.DomainException;
+import br.edu.com.tg.manager.core.port.csvreader.CsvReaderPort;
 import br.edu.com.tg.manager.core.port.repository.CourseRepository;
 import br.edu.com.tg.manager.core.port.repository.StudentGroupRepository;
 import br.edu.com.tg.manager.core.port.repository.StudentRepository;
-import br.edu.com.tg.manager.core.port.CsvReaderPort;
 import br.edu.com.tg.manager.core.usecase.ImportStudentsFromCsvUseCase;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
